# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","description":"presence of request param is due to implicit requirement of slowapi","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/carparks/nearby":{"get":{"summary":"Nearby Carparks","description":"fastapi depends auto verifies the validity of api_key","operationId":"nearby_carparks_carparks_nearby_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","exclusiveMaximum":90,"exclusiveMinimum":-90,"title":"Lat"}},{"name":"lng","in":"query","required":true,"schema":{"type":"number","exclusiveMaximum":180,"exclusiveMinimum":-180,"title":"Lng"}},{"name":"radius_km","in":"query","required":false,"schema":{"type":"number","exclusiveMinimum":0,"default":5,"title":"Radius Km"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CarparkRelative"},"title":"Response Nearby Carparks Carparks Nearby Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/carparks/{facility_id}":{"get":{"summary":"Carpark Detail","operationId":"carpark_detail_carparks__facility_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"facility_id","in":"path","required":true,"schema":{"type":"string","title":"Facility Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarparkDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CarparkDetail":{"properties":{"facility_id":{"type":"string","title":"Facility Id"},"name":{"type":"string","title":"Name"},"total_spots":{"type":"integer","title":"Total Spots"},"available_spots":{"type":"integer","title":"Available Spots"},"status":{"type":"string","title":"Status"},"last_updated":{"type":"string","format":"date-time","title":"Last Updated"}},"type":"object","required":["facility_id","name","total_spots","available_spots","status","last_updated"],"title":"CarparkDetail"},"CarparkRelative":{"properties":{"facility_id":{"type":"string","title":"Facility Id"},"name":{"type":"string","title":"Name"},"distance_km":{"type":"number","title":"Distance Km"}},"type":"object","required":["facility_id","name","distance_km"],"title":"CarparkRelative"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}